First Update
There is a clock on the screen. IF you click on it , it will turn into military clock and viceversa. if you click on the edit buttton above the clock you can select your preferred background image. here goes some server side work so ive just hard coded the change when action is performed. nontheless the actions are working correctly on my laptop. savy